fix ordenamiento y desacoplamiento de componentes en categoryScreen
