+++ app/build.gradle
@@ -50,6 +50,7 @@ android {
 
 dependencies {
     implementation 'com.google.android.gms:play-services-mlkit-text-recognition-common:19.0.0'
+    implementation 'com.android.support:support-annotations:28.0.0'
     def nav_version = "2.7.1"
     def hilt_version = "2.47"
     def retrofit_version = "2.9.0"
@@ -88,10 +89,10 @@ dependencies {
     debugImplementation 'androidx.compose.ui:ui-test-manifest'
 
     // Coroutines
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.2'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.2'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.1'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.1'
 
     // LyfeCycle Scopes
-    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.3.1'
-    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.3.1'
+    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.6.2'
+    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.6.2'
 }
